requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.4
requests==2.31.0
pandas==2.1.3
numpy==1.24.4
matplotlib==3.8.2
seaborn==0.13.0
textblob==0.17.1
nltk==3.8.1
wordcloud==1.9.2
python-dotenv==1.0.0
pyyaml==6.0.1
jsonschema==4.20.0
